Keep rating priority single-file compatible
