+++ smartgallery.py
@@ -31,7 +31,6 @@ from tqdm import tqdm
 import threading
 import uuid
 import socket
-from rating_priority import prioritize_personal_unrated, resolve_rating_client_uuid
 # Try to import tkinter for GUI dialogs, but make it optional for Docker/headless environments
 try:
     import tkinter as tk
@@ -690,6 +689,22 @@ app.secret_key = SECRET_KEY
 gallery_view_cache = []
 folder_config_cache = None
 FFPROBE_EXECUTABLE_PATH = None
+
+
+def resolve_rating_client_uuid(session_user_id, force_login, exhibition_mode):
+    """Return the identity used by ``file_ratings`` for the current request."""
+    if session_user_id is not None and str(session_user_id):
+        return str(session_user_id)
+    if not force_login and not exhibition_mode:
+        return "admin"
+    return ""
+
+
+def prioritize_personal_unrated(files):
+    """Stable-partition files so unrated items precede personally rated items."""
+    for index, item in enumerate(files):
+        item["review_sort_index"] = index
+    return sorted(files, key=lambda item: bool(item.get("my_rating")))
 
 
 def current_rating_client_uuid():
